fix: refactor extract middlewares form routes.ts
